Move ParsedKey to parsed_key.rs
